Improve `pico_flash` task
This set of changes will use `picotool` to reset the rp2040 device if a serial
monitor is attached. A warning will also be issued notifying the user that
faster flashing times can be achieved by disconnecting the serial monitor
application before using the `atomvm pico_flash` task, as this also requires an
extra delay before it can be put back into `BOOTSEL` mode.

Signed-off-by: Winford <winford@object.stream>