Enhancements to pico_flash task
Improves compatibility with RP2350 devices, detecting path and reset devices correctly so setting
explicit path to the mount point is no longer necessary for these devices. The provider also now
confirms that the target is an RP2 device before resetting to `BOOTSEL` mode, preventing
interference with non RP2 devices that may be attached to the host system.

Closes #44

Signed-off-by: Winford <winford@object.stream>