Make atomvm_pico_flash_provider easier to test
Add some more verbose logging (when in test mode) to give test the ability to parse parameters sent
to picotool. Reduces times waiting for mount points, since we are not waiting for real hardware to
settle. Allows the use of a mock esptool for testing under CI or when no RP2 hardware is present.

Signed-off-by: Winford <winford@object.stream>